refactor: ongoing contract slider
